requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
bcrypt==3.2.0
fhir.resources==6.1.0
hl7==0.4.2
